Move is_safe_url/get_next_url into apps.common, and use that.
